Get suggested follow artists on password view load
